requirements.txt
langchain-text-splitters>=0.0.1
langchain-ollama>=0.0.1
langchain-community>=0.0.10
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
streamlit>=1.32.0
langchain-groq>=0.0.3
pypdf>=3.0.0
python-dotenv>=0.19.0
langchain>=0.1.0
groq>=0.4.0
openai>=1.12.0
langchain-openai>=0.0.5
langchain-core==0.1.10 